Move tests to scripts/tests and change print to logger.info
