+++ scripts/migration/migrate_none_as_email_verification.py
@@ -3,13 +3,12 @@
 
 import logging
 import sys
-from tests.base import OsfTestCase
-from tests.factories import UserFactory
 from modularodm import Q
 from nose.tools import *
 from website import models
 from website.app import init_app
 from scripts import utils as scripts_utils
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
@@ -28,22 +27,10 @@ def main():
         if not dry_run:
             user.save()
 
-    print('{} users migrated'.format(count))
+    logger.info('Done with {} users migrated'.format(count))
 
 def get_users_with_none_in_email_verifications():
-    return models.User.find( Q('email_verifications', 'eq', None))
-
-class TestMigrateDates(OsfTestCase):
-    def setUp(self):
-        super(TestMigrateDates, self).setUp()
-        self.user1 = UserFactory(email_verfications=None)
-        self.user2 = UserFactory(email_verfications={})
-
-    def test_migrate_none_as_email(self):
-        main()
-        assert_equal(self.user1.email_verifications, {})
-        assert_not_equal(self.user2.email_verifications, None)
-
+    return models.User.find(Q('email_verifications', 'eq', None))
 
 if __name__ == '__main__':
-        main()
+    main()
